corrected: missing references to instance and filesystem. (ejbPostCreate not called because of the wrong method signature!)

+++ dcm4chee/dcm4chee-arc/trunk/dcm4jboss-ejb/src/java/org/dcm4chex/archive/ejb/entity/PrivateFileBean.java
@@ -84,7 +84,7 @@ public abstract class PrivateFileBean implements EntityBean {
         return null;
     }
 
-    public void ejbPostCreate(String path, String tsuid, int size, byte[] md5,
+    public void ejbPostCreate(String path, String tsuid, long size, byte[] md5,
     		int status, PrivateInstanceLocal instance, FileSystemLocal filesystem)
             throws CreateException {
         setInstance(instance);
